// gascd_frontend/src/data/interfaces/Locations.ts
export interface Locations {
  provider_location_id: string;
  provider_location_name: string;
  provider_id: string;
  provider_name: string;
  la_code: string;
  la_name: string;
  region_code: string;
  region_name: string;
  country_code: string;
  country_name: string;
  load_date_time: string;
}